edited the function

+++ add_func.c
@@ -8,7 +8,7 @@
 void print_integer(int *count, va_list args)
 {
 	int n = va_arg(args, int);
-	unsigned int num = n;
+	unsigned int num = (unsigned int) (n);
 	unsigned int divisor;
 
 	if (num < 0)
